allow individual extensions to be loaded
fixes:
    gems/minitest-around-0.0.1/lib/minitest/around.rb:3:in
    `<top (required)>': uninitialized constant MiniTest::Unit (NameError)
