+++ lib/minitest/around/spec.rb
@@ -1,0 +1,11 @@
+class MiniTest::Spec
+  def self.around(&outer)
+    define_method(:around) do |&inner|
+      if outer.arity == 1
+        outer.call(inner)
+      else
+        inner.call *outer.call
+      end
+    end
+  end
+end
